requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.0
llama_index==0.11.16
llama_parse==0.5.7
Markdown==3.7
nest_asyncio==1.6.0
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
uvicorn==0.31.0
python-multipart==0.0.12
pipreqs==0.5.0
alembic==1.13.3
pillow==10.4.0
pypdfium2==4.30.0
scikit-learn==1.6.0
sentence-transformers==3.3.1
joblib==1.4.2
numpy==1.26.4
openai==1.59.9
regex==2024.9.11
pandas==2.2.3
openpyxl==3.1.5
pgvector==0.4.0
